+++ apps/client/src/app/components/investment-chart/investment-chart.component.ts
@@ -16,6 +16,7 @@ import { LineChartItem } from '@ghostfolio/common/interfaces';
 import { InvestmentItem } from '@ghostfolio/common/interfaces/investment-item.interface';
 import { ColorScheme, GroupBy } from '@ghostfolio/common/types';
 
+import { CommonModule } from '@angular/common';
 import {
   ChangeDetectionStrategy,
   Component,
@@ -24,7 +25,6 @@ import {
   OnDestroy,
   ViewChild
 } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import {
   BarController,
   BarElement,
@@ -44,13 +44,13 @@ import { isAfter } from 'date-fns';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 
 @Component({
-  selector: 'gf-investment-chart',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: './investment-chart.component.html',
+  imports: [CommonModule, NgxSkeletonLoaderModule],
+  selector: 'gf-investment-chart',
   styleUrls: ['./investment-chart.component.scss'],
-  imports: [CommonModule, NgxSkeletonLoaderModule]
+  templateUrl: './investment-chart.component.html'
 })
-export class InvestmentChartComponent implements OnChanges, OnDestroy {
+export class GfInvestmentChartComponent implements OnChanges, OnDestroy {
   @Input() benchmarkDataItems: InvestmentItem[] = [];
   @Input() benchmarkDataLabel = '';
   @Input() colorScheme: ColorScheme;