+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- Refactored the investment chart component to standalone
 
 ### Changed
 
 - Refactored the header component to standalone
+- Refactored the investment chart component to standalone
 
 ## 2.197.0 - 2025-09-07